+++ src/main/java/org/kdb/inside/brains/ide/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains only files related to IDE implementation.
+ * <p>
+ * Code from this package can be used only in plugin-modules-java.xml file or the plugin won't work on other platforms.
+ */
+package org.kdb.inside.brains.ide;