+++ src/main/java/org/kdb/inside/brains/lang/template/QTemplateContextType.java
@@ -1,12 +1,12 @@
-package org.kdb.inside.brains.ide.template;
+package org.kdb.inside.brains.lang.template;
 
 import com.intellij.codeInsight.template.TemplateActionContext;
 import com.intellij.codeInsight.template.TemplateContextType;
 import org.jetbrains.annotations.NotNull;
 import org.kdb.inside.brains.QFileType;
 
-public class KdbTemplateContextType extends TemplateContextType {
-    protected KdbTemplateContextType() {
+public class QTemplateContextType extends TemplateContextType {
+    protected QTemplateContextType() {
         super("KDB+ Q");
     }