+++ src/main/java/org/kdb/inside/brains/psi/QIconProvider.java
@@ -3,11 +3,9 @@ package org.kdb.inside.brains.psi;
 import com.intellij.ide.IconProvider;
 import com.intellij.openapi.project.DumbAware;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.util.PsiUtil;
 import com.intellij.ui.IconManager;
 import com.intellij.ui.icons.RowIcon;
 import com.intellij.util.BitUtil;
-import com.intellij.util.VisibilityIcons;
 import icons.KdbIcons;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -33,15 +31,13 @@ public class QIconProvider extends IconProvider implements DumbAware {
             return KdbIcons.Node.Context;
         } else if (element instanceof QSymbol) {
             return KdbIcons.Node.Symbol;
-        } else if (element instanceof QTableColumn) {
-            return getColumnIcon((QTableColumn) element);
+        } else if (element instanceof QTableColumn tbl) {
+            return getColumnIcon(tbl);
         } else if (element instanceof QLambdaExpr) {
             return KdbIcons.Node.Lambda;
-        } else if (element instanceof QAssignmentExpr) {
-            final QAssignmentExpr assignment = (QAssignmentExpr) element;
+        } else if (element instanceof QAssignmentExpr assignment) {
             return getAssignmentIcon(assignment, visibility);
-        } else if (element instanceof QVarDeclaration) {
-            final QVarDeclaration declaration = (QVarDeclaration) element;
+        } else if (element instanceof QVarDeclaration declaration) {
             return getIcon(declaration.getParent(), flags);
         }
         return null;
@@ -56,14 +52,14 @@ public class QIconProvider extends IconProvider implements DumbAware {
         Icon i = getExpressionIcon(expression);
         if (visibility) {
             final RowIcon icon = IconManager.getInstance().createLayeredIcon(assignment, i, 0);
-            if (QPsiUtil.isGlobalDeclaration(assignment)) {
-                VisibilityIcons.setVisibilityIcon(PsiUtil.ACCESS_LEVEL_PUBLIC, icon);
-            } else {
-                VisibilityIcons.setVisibilityIcon(PsiUtil.ACCESS_LEVEL_PRIVATE, icon);
-            }
+            icon.setIcon(getVisibilityIcon(assignment), 1);
             return icon;
         }
         return i;
+    }
+
+    private Icon getVisibilityIcon(QAssignmentExpr assignment) {
+        return QPsiUtil.isGlobalDeclaration(assignment) ? KdbIcons.Node.PublicItem : KdbIcons.Node.PrivateItem;
     }
 
     private Icon getExpressionIcon(QExpression expression) {