+++ src/main/java/org/kdb/inside/brains/view/inspector/InspectorToolWindowFactory.java
@@ -1,23 +1,15 @@
 package org.kdb.inside.brains.view.inspector;
 
-import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowContentUiType;
-import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.openapi.wm.ex.ToolWindowEx;
 import org.jetbrains.annotations.NotNull;
+import org.kdb.inside.brains.view.KdbToolWindowFactory;
 
-public class InspectorToolWindowFactory implements ToolWindowFactory, DumbAware {
+public class InspectorToolWindowFactory implements KdbToolWindowFactory {
     @Override
-    public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
-        if (!(toolWindow instanceof ToolWindowEx)) {
-            return;
-        }
-
-        final ToolWindowEx tw = (ToolWindowEx) toolWindow;
-        tw.setDefaultContentUiType(ToolWindowContentUiType.TABBED);
-
-        project.getService(InspectorToolWindow.class).initToolWindow(tw);
+    public void createToolWindowContentEx(@NotNull Project project, @NotNull ToolWindowEx toolWindow) {
+        toolWindow.setDefaultContentUiType(ToolWindowContentUiType.TABBED);
+        project.getService(InspectorToolWindow.class).initToolWindow(toolWindow);
     }
 }