+++ src/main/java/org/kdb/inside/brains/view/treeview/InstancesToolWindowFactory.java
@@ -1,29 +1,15 @@
 package org.kdb.inside.brains.view.treeview;
 
-import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowContentUiType;
-import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.openapi.wm.ex.ToolWindowEx;
 import org.jetbrains.annotations.NotNull;
+import org.kdb.inside.brains.view.KdbToolWindowFactory;
 
-public class InstancesToolWindowFactory implements ToolWindowFactory, DumbAware {
+public class InstancesToolWindowFactory implements KdbToolWindowFactory {
     @Override
-    public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
-        if (!(toolWindow instanceof ToolWindowEx)) {
-            return;
-        }
-
-        final ToolWindowEx tw = (ToolWindowEx) toolWindow;
-        tw.setDefaultContentUiType(ToolWindowContentUiType.TABBED);
-
-        project.getService(InstancesToolWindow.class).initToolWindow(tw);
-    }
-
-    @Override
-    public boolean shouldBeAvailable(@NotNull Project project) {
-        // TODO: Any way to check that it's KDB project here?
-        return true;
+    public void createToolWindowContentEx(@NotNull Project project, @NotNull ToolWindowEx toolWindow) {
+        toolWindow.setDefaultContentUiType(ToolWindowContentUiType.TABBED);
+        project.getService(InstancesToolWindow.class).initToolWindow(toolWindow);
     }
 }